Fix conversion problems
Fix computing depth map scale which was wrongly doing integer arithmetic
to compute a floating point value. Modify camera scale methods to take a
double rather than a float, as the internal math is done on doubles
anyway. Fix some other conversion issues.

+++ gui/vtkMaptkCamera.h
@@ -59,9 +59,9 @@ public:
   // specified depth
   kwiver::vital::vector_3d UnprojectPoint(double point[2], double depth);
 
-  void ScaleK(float factor);
+  void ScaleK(double factor);
 
-  vtkSmartPointer<vtkMaptkCamera> ScaledK(float factor);
+  vtkSmartPointer<vtkMaptkCamera> ScaledK(double factor);
 
   // Description:
   // Update self (the VTK camera) based on the maptk camera and
